backend/requirements.txt
# =========================
# BARD BACKEND DEPENDENCIES
# =========================

# ----- Django Core -----
Django>=5.0,<5.1
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=23.5,<24.0
django-environ>=0.11,<1.0

# ----- Database -----
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.1,<3.0

# ----- Authentication -----
djangorestframework-simplejwt>=5.3,<6.0
django-allauth>=0.60,<1.0

# ----- Async Tasks -----
celery>=5.3,<6.0
django-celery-beat>=2.5,<3.0
django-celery-results>=2.5,<3.0
redis>=5.0,<6.0

# ----- Web Scraping (Spider Module) -----
beautifulsoup4>=4.12,<5.0
requests>=2.31,<3.0
lxml>=5.0,<6.0
playwright>=1.40,<2.0
httpx>=0.26,<1.0

# ----- Data Processing -----
pandas>=2.1,<3.0
numpy>=1.26,<2.0

# ----- Rich Text & Content -----
django-ckeditor>=6.7,<7.0
Pillow>=10.1,<11.0
django-storages>=1.14,<2.0
boto3>=1.34,<2.0

# ----- Email -----
django-anymail>=10.2,<11.0

# ----- API Documentation -----
drf-spectacular>=0.27,<1.0

# ----- Caching -----
django-redis>=5.4,<6.0

# ----- Production Server -----
gunicorn>=21.2,<22.0
uvicorn>=0.25,<1.0
whitenoise>=6.6,<7.0

# ----- Monitoring & Logging -----
sentry-sdk>=1.39,<2.0
django-structlog>=6.0,<7.0

# ----- Development & Testing -----
pytest>=7.4,<8.0
pytest-django>=4.7,<5.0
pytest-cov>=4.1,<5.0
factory-boy>=3.3,<4.0
faker>=22.0,<23.0
black>=23.12,<24.0
isort>=5.13,<6.0
flake8>=7.0,<8.0
mypy>=1.8,<2.0
django-stubs>=4.2,<5.0
